refactor(frontend): extract shared CardSearchInput + CardFetchMoreError from cards clients
